+++ langchain4j-mcp/src/test/java/dev/langchain4j/mcp/client/integration/McpToolsTestBase.java
@@ -14,28 +14,25 @@ import dev.langchain4j.model.chat.request.json.JsonIntegerSchema;
 import dev.langchain4j.model.chat.request.json.JsonStringSchema;
 import dev.langchain4j.model.openai.OpenAiChatModel;
 import dev.langchain4j.service.AiServices;
+import dev.langchain4j.service.AbstractAiServicesWithToolErrorHandlerTest;
+import dev.langchain4j.service.tool.ToolArgumentParsingException;
 import dev.langchain4j.service.tool.ToolExecutionException;
 import dev.langchain4j.service.tool.ToolExecutor;
-import dev.langchain4j.service.tool.ToolProvider;
 import dev.langchain4j.service.tool.ToolProviderResult;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public abstract class McpToolsTestBase {
+public abstract class McpToolsTestBase extends AbstractAiServicesWithToolErrorHandlerTest {
 
     static McpClient mcpClient;
-
-    private static final Logger log = LoggerFactory.getLogger(McpToolsTestBase.class);
 
     @Test
     public void verifyToolSpecifications() {
         ToolProviderResult toolProviderResult = obtainTools();
 
         Map<ToolSpecification, ToolExecutor> tools = toolProviderResult.tools();
-        assertThat(tools).hasSize(7);
+        assertThat(tools).hasSize(9);
 
         ToolSpecification echoString = toolProviderResult.toolSpecificationByName("echoString");
         assertThat(echoString.description()).isEqualTo("Echoes a string");
@@ -104,7 +101,7 @@ public abstract class McpToolsTestBase {
                 .arguments("{\"input\": 1}")
                 .build();
         assertThatThrownBy(() -> executor.execute(toolExecutionRequest, null))
-                .isExactlyInstanceOf(ToolExecutionException.class)
+                .isExactlyInstanceOf(ToolArgumentParsingException.class)
                 .hasMessage("Invalid tool name: THIS-TOOL-DOES-NOT-EXIST")
                 .hasFieldOrPropertyWithValue("errorCode", -32602);
     }
@@ -160,10 +157,8 @@ public abstract class McpToolsTestBase {
     @Test
     @EnabledIfEnvironmentVariable(named = "OPENAI_API_KEY", matches = ".+")
     public void executeToolWithUntypedArrayParameter() {
-        ToolProvider toolProvider =
-                McpToolProvider.builder().mcpClients(mcpClient).build();
         ChatService service = AiServices.builder(ChatService.class)
-                .toolProvider(toolProvider)
+                .toolProvider(createMcpToolProvider())
                 .chatMemory(MessageWindowChatMemory.withMaxMessages(10))
                 .chatModel(chatModel)
                 .build();
@@ -177,6 +172,22 @@ public abstract class McpToolsTestBase {
     }
 
     ToolProviderResult obtainTools() {
-        return McpToolProvider.builder().mcpClients(mcpClient).build().provideTools(null);
+        return createMcpToolProvider().provideTools(null);
+    }
+
+    @Override
+    protected void configureGetWeatherThrowingExceptionTool(RuntimeException ignored, AiServices<?> aiServiceBuilder) {
+        aiServiceBuilder.toolProvider(createMcpToolProvider());
+    }
+
+    @Override
+    protected void configureGetWeatherTool(AiServices<?> aiServiceBuilder) {
+        aiServiceBuilder.toolProvider(createMcpToolProvider());
+    }
+
+    private static McpToolProvider createMcpToolProvider() {
+        return McpToolProvider.builder()
+                .mcpClients(mcpClient)
+                .build();
     }
 }
