+++ langchain4j/src/test/java/dev/langchain4j/service/AiServicesWithToolsIT.java
@@ -6,20 +6,16 @@ import static dev.langchain4j.service.AiServicesWithToolsIT.TransactionService.E
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonMap;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
 import static org.assertj.core.data.MapEntry.entry;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 
-import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -34,7 +30,6 @@ import dev.langchain4j.data.message.UserMessage;
 import dev.langchain4j.memory.ChatMemory;
 import dev.langchain4j.memory.chat.MessageWindowChatMemory;
 import dev.langchain4j.model.chat.ChatModel;
-import dev.langchain4j.model.chat.mock.ChatModelMock;
 import dev.langchain4j.model.chat.request.ChatRequest;
 import dev.langchain4j.model.chat.request.json.JsonArraySchema;
 import dev.langchain4j.model.chat.request.json.JsonIntegerSchema;
@@ -43,11 +38,7 @@ import dev.langchain4j.model.chat.request.json.JsonStringSchema;
 import dev.langchain4j.model.chat.response.ChatResponse;
 import dev.langchain4j.model.openai.OpenAiChatModel;
 import dev.langchain4j.model.output.TokenUsage;
-import dev.langchain4j.service.tool.ToolArgumentParsingException;
-import dev.langchain4j.service.tool.ToolErrorHandler;
-import dev.langchain4j.service.tool.ToolErrorHandlerResult;
 import dev.langchain4j.service.tool.ToolExecution;
-import dev.langchain4j.service.tool.ToolExecutionException;
 import dev.langchain4j.service.tool.ToolExecutor;
 import dev.langchain4j.service.tool.ToolProvider;
 import dev.langchain4j.service.tool.ToolProviderResult;
@@ -66,7 +57,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
@@ -1094,344 +1084,5 @@ class AiServicesWithToolsIT {
         assertThat(messages.get(1)).isInstanceOf(AiMessage.class); // ai message to invoke the tool
         assertThat(messages.get(2)).isInstanceOf(ToolExecutionResultMessage.class); // tool response
         assertThat(messages.get(3)).isInstanceOf(AiMessage.class); // final ai message
-    }
-
-    // Error Handling: Tool Error
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_propagate_error_message_thrown_from_tool_to_LLM_by_default(boolean executeToolsConcurrently) {
-
-        // given
-        String errorMessage = "Weather service is unavailable";
-
-        class FailingTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                throw new RuntimeException(errorMessage);
-            }
-        }
-
-        ChatModel spyModel = spy(models().findFirst().get());
-
-        FailingTool spyTool = spy(new FailingTool());
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // when
-        assistant.chat("What is the weather in Munich?");
-
-        // then
-        verify(spyTool).getWeather("Munich");
-        verifyNoMoreInteractions(spyTool);
-
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 1));
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 3
-                && chatRequest.messages().get(2) instanceof ToolExecutionResultMessage toolResult
-                && toolResult.text().equals(errorMessage)));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_customize_error_returned_from_tool_before_sending_to_LLM(boolean executeToolsConcurrently) {
-
-        // given
-        RuntimeException toolError = new RuntimeException("Can't connect to the reliable-weather.com");
-
-        class FailingTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                throw toolError;
-            }
-        }
-
-        String customizedErrorMessage = "Weather service is unavailable";
-
-        ToolErrorHandler toolErrorHandler = (error, context) -> {
-            assertThat(error)
-                    .isExactlyInstanceOf(ToolExecutionException.class)
-                    .hasCause(toolError);
-            assertThat(context.toolExecutionRequest().name()).isEqualTo("getWeather");
-            assertThat(context.toolExecutionRequest().arguments()).contains("Munich");
-            assertThat(context.memoryId()).isEqualTo("default");
-
-            return ToolErrorHandlerResult.from(customizedErrorMessage);
-        };
-
-        ChatModel spyModel = spy(models().findFirst().get());
-
-        FailingTool spyTool = spy(new FailingTool());
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool)
-                .toolErrorHandler(toolErrorHandler);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // when
-        assistant.chat("What is the weather in Munich?");
-
-        // then
-        verify(spyTool).getWeather("Munich");
-        verifyNoMoreInteractions(spyTool);
-
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 1));
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 3
-                && chatRequest.messages().get(2) instanceof ToolExecutionResultMessage toolResult
-                && toolResult.text().equals(customizedErrorMessage)));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_fail_when_tool_throws_error(boolean executeToolsConcurrently) {
-
-        // given
-        RuntimeException toolError = new RuntimeException("Weather service is unavailable");
-
-        class FailingTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                throw toolError;
-            }
-        }
-
-        ToolErrorHandler toolErrorHandler = (error, context) -> {
-            assertThat(error)
-                    .isExactlyInstanceOf(ToolExecutionException.class)
-                    .hasCause(toolError);
-            assertThat(context.toolExecutionRequest().name()).isEqualTo("getWeather");
-            assertThat(context.toolExecutionRequest().arguments()).contains("Munich");
-            assertThat(context.memoryId()).isEqualTo("default");
-
-            throw toolError;
-        };
-
-        ChatModel spyModel = spy(models().findFirst().get());
-
-        FailingTool spyTool = spy(new FailingTool());
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool)
-                .toolErrorHandler(toolErrorHandler);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // asking for 2 tools to make sure tools will be executed concurrently using executor
-        String userMessage = "What is the weather in Munich and Paris?";
-
-        // when
-        assertThatThrownBy(() -> assistant.chat(userMessage))
-                .isSameAs(toolError);
-
-        // then
-        verify(spyTool).getWeather("Munich");
-        if (executeToolsConcurrently) {
-            // when executed sequentially, second tool is not executed,
-            // the whole AI Service fails after the first tool fails
-            verify(spyTool).getWeather("Paris");
-        }
-        verifyNoMoreInteractions(spyTool);
-
-        verify(spyModel).chat(any(ChatRequest.class));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    // Error Handling: Argument Error
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_fail_when_cannot_parse_tool_arguments_by_default(boolean executeToolsConcurrently) {
-
-        // given
-        ToolExecutionRequest toolExecutionRequest = ToolExecutionRequest.builder()
-                .name("getWeather")
-                .arguments("{ invalid json }")
-                .build();
-
-        ChatModel spyModel = spy(ChatModelMock.thatAlwaysResponds(AiMessage.from(toolExecutionRequest)));
-
-        class WeatherTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                return "sunny";
-            }
-        }
-
-        WeatherTool spyTool = spy(new WeatherTool());
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // when
-        assertThatThrownBy(() -> assistant.chat("What is the weather in Munich?"))
-                .isExactlyInstanceOf(RuntimeException.class)
-                .hasCauseExactlyInstanceOf(JsonParseException.class)
-                .hasMessageContaining("Unexpected character");
-
-        // then
-        verifyNoInteractions(spyTool);
-
-        verify(spyModel).chat(any(ChatRequest.class));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_customize_argument_parsing_error_before_sending_to_LLM(boolean executeToolsConcurrently) {
-
-        // given
-        ToolExecutionRequest toolExecutionRequest1 = ToolExecutionRequest.builder()
-                .name("getWeather")
-                .arguments("{ invalid json }")
-                .build();
-
-        ToolExecutionRequest toolExecutionRequest2 = ToolExecutionRequest.builder()
-                .name("getWeather")
-                .arguments("{\"arg0\":\"Munich\"}")
-                .build();
-
-        ChatModel spyModel = spy(ChatModelMock.thatAlwaysResponds(
-                AiMessage.from(toolExecutionRequest1),
-                AiMessage.from(toolExecutionRequest2),
-                AiMessage.from("sunny")
-        ));
-
-        class WeatherTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                return "sunny";
-            }
-        }
-
-        WeatherTool spyTool = spy(new WeatherTool());
-
-        String customizedErrorMessage = "Invalid JSON, try again";
-
-        ToolErrorHandler toolErrorHandler = (error, context) -> {
-            assertThat(error)
-                    .isExactlyInstanceOf(ToolArgumentParsingException.class)
-                    .hasCauseExactlyInstanceOf(JsonParseException.class)
-                    .hasMessageContaining("Unexpected character");
-            assertThat(context.toolExecutionRequest()).isEqualTo(toolExecutionRequest1);
-            assertThat(context.memoryId()).isEqualTo("default");
-
-            return ToolErrorHandlerResult.from(customizedErrorMessage);
-        };
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool)
-                .toolErrorHandler(toolErrorHandler);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // when
-        assistant.chat("What is the weather in Munich?");
-
-        // then
-        verify(spyTool).getWeather("Munich");
-        verifyNoMoreInteractions(spyTool);
-
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 1));
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 3
-                && chatRequest.messages().get(2) instanceof ToolExecutionResultMessage toolResult
-                && toolResult.text().equals(customizedErrorMessage)));
-        verify(spyModel).chat(argThat((ChatRequest chatRequest) -> chatRequest.messages().size() == 5));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    @ParameterizedTest
-    @ValueSource(booleans = {false, true})
-    void should_fail_with_custom_exception_when_tool_arguments_cannot_be_parsed(boolean executeToolsConcurrently) {
-
-        // given
-        ToolExecutionRequest toolExecutionRequest = ToolExecutionRequest.builder()
-                .name("getWeather")
-                .arguments("{ invalid json }")
-                .build();
-
-        ChatModel spyModel = spy(ChatModelMock.thatAlwaysResponds(AiMessage.from(toolExecutionRequest)));
-
-        class WeatherTool {
-
-            @Tool
-            String getWeather(String ignored) {
-                return "sunny";
-            }
-        }
-
-        WeatherTool spyTool = spy(new WeatherTool());
-
-        RuntimeException customException = new RuntimeException("Can't parse JSON arguments");
-
-        ToolErrorHandler toolErrorHandler = (error, context) -> {
-            assertThat(error)
-                    .isExactlyInstanceOf(ToolArgumentParsingException.class)
-                    .hasCauseExactlyInstanceOf(JsonParseException.class)
-                    .hasMessageContaining("Unexpected character");
-            assertThat(context.toolExecutionRequest()).isEqualTo(toolExecutionRequest);
-            assertThat(context.memoryId()).isEqualTo("default");
-
-            throw customException;
-        };
-
-        AiServices<Assistant> assistantBuilder = AiServices.builder(Assistant.class)
-                .chatModel(spyModel)
-                .tools(spyTool)
-                .toolErrorHandler(toolErrorHandler);
-        if (executeToolsConcurrently) {
-            assistantBuilder.executeToolsConcurrently();
-        }
-        Assistant assistant = assistantBuilder.build();
-
-        // when
-        assertThatThrownBy(() -> assistant.chat("What is the weather in Munich?"))
-                .isSameAs(customException);
-
-        // then
-        verifyNoInteractions(spyTool);
-
-        verify(spyModel).chat(any(ChatRequest.class));
-        ignoreOtherInteractions(spyModel);
-        verifyNoMoreInteractions(spyModel);
-    }
-
-    private static void ignoreOtherInteractions(ChatModel model) {
-        verify(model, atLeast(0)).doChat(any());
-        verify(model, atLeast(0)).defaultRequestParameters();
-        verify(model, atLeast(0)).listeners();
-        verify(model, atLeast(0)).provider();
-        verify(model, atLeast(0)).supportedCapabilities();
     }
 }
